C9.1: drop migration, repository interface, follow/unfollow use cases
