+++ apps/mobile/src/core/query/keys/positions.keys.ts
@@ -1,5 +1,5 @@
 import { createQueryKeys } from "@lukemorales/query-key-factory";
-import type { TPositionsListParams } from "positions/data/positions.repository";
+import type { TPositionsListParams } from "positions/data/gateway-positions.repository";
 
 export const positionsKeys = createQueryKeys("positions", {
   list: (params: Omit<TPositionsListParams, "cursor">) => ({
