+++ apps/mobile/src/positions/di/tokens.ts
@@ -1,2 +1,1 @@
-export const POSITIONS_REPOSITORY = Symbol("POSITIONS_REPOSITORY");
 export const FOLLOWING_REPOSITORY = Symbol("FOLLOWING_REPOSITORY");
