+++ apps/mobile/src/positions/presentation/hooks/useFollow.ts
@@ -1,8 +1,6 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { container } from "core/di/container";
 import { positionsKeys } from "core/query/keys/positions.keys";
-import { FollowPositionUseCase } from "positions/application/usecase/follow-position.usecase";
-import { UnfollowPositionUseCase } from "positions/application/usecase/unfollow-position.usecase";
 import { FollowingStore } from "positions/presentation/stores/following.store";
 
 export function useFollow(ref: string) {
@@ -10,9 +8,8 @@ export function useFollow(ref: string) {
   const store = container.resolve(FollowingStore);
   return useMutation({
     mutationFn: async (next: boolean) => {
-      if (next) await container.resolve(FollowPositionUseCase).execute(ref);
-      else await container.resolve(UnfollowPositionUseCase).execute(ref);
-      store.toggle(ref);
+      if (next) store.follow(ref);
+      else store.unfollow(ref);
     },
     onSuccess: () => qc.invalidateQueries({ queryKey: positionsKeys._def }),
   });
