+++ apps/mobile/src/positions/presentation/hooks/usePositionByRefQuery.ts
@@ -1,11 +1,10 @@
 import { useQuery } from "@tanstack/react-query";
 import { container } from "core/di/container";
 import { positionsKeys } from "core/query/keys/positions.keys";
-import type { IPositionsRepository } from "positions/data/positions.repository";
-import { POSITIONS_REPOSITORY } from "positions/di/tokens";
+import { GatewayPositionsRepository } from "positions/data/gateway-positions.repository";
 
 export function usePositionByRefQuery(ref: string) {
-  const repo = container.resolve<IPositionsRepository>(POSITIONS_REPOSITORY);
+  const repo = container.resolve(GatewayPositionsRepository);
   return useQuery({
     queryKey: positionsKeys.byRef(ref).queryKey,
     queryFn: () => repo.getByRef(ref),
